Fix mummerplot when ^perl@5.20: (#4348)
* Fix mummerplot when ^perl@5.20:

Calling defined() on a hash has been deprecated for ages.  It became
an error in perl@5.20.  If we're building with a perl where it's
illegal, we should fix it.

* Simplify call to filter_file

Treat the first arg to filter_file as a string rather than
a regex, so that we don't have to figure out the escapes.

* Patch mummerplot for *any* perl version.