[WIP]: Sambamba new package (#5304)
* First draft package for sambamba

Depends on the ldc package (#5279) that alalazo's working on.

It requires a patch to the Makefile because one of ldc's library
functions mentions dlopen.  I *think* that ldc should be handling
this.  See the discusssion in #5279.

It also requires a "static" ldc.  Again, see #5279.

Finally, it needs "undeaD" as a resource, but I'm guessing at the
appropriate release.  See
https://github.com/lomereiter/sambamba/issues/312.

* alalazo renamed the shared_libs variant, keep up...