libc: Return empty libs (#20338)
The C-Library for the current compiler should already be used by the compiler. So there is no point in returning any libs for this package.

Without this patch: if one uses this as an external package (as intended), then this will can inject system library paths into the build process at the wrong place.