eospac: add version 6.4.2beta (#20681)
* eospac: add version 6.4.2beta

* eospac: clarify EOSPAC "beta" versions

Compared to 6.4.1, EOSPAC 6.4.2beta contains only one change, a fix
for an inability to read some SESAME files in ASCII format.  From the
release announcement,

    EOSPAC 6.4.2beta has been released for general use as the latest
    (i.e., eospac6-latest) versions.  This is a small patch to the
    previously-released version 6.4.1, which was requested by an
    affected user.

But the "beta" label can cause confusion, especially when a beta
version is the new preferred version, as is the case here.  As
suggested by reviewers, add a comment clarifying EOSPAC's use of
"beta".