mumps: Fix for problematic src/makefile patch (#20590) (#20758)
* mumps: Fix for problematic src/makefile patch (#20590)

Minor change in src/Makefile between 5.2.0 and 5.3.3 causing patch to
break.  Split into 2 patchfiles

* mumps: Additional patch for fixing #20590

This is to fix issue wherein build fails on Ubuntu due to undefined
symbols, despite symbols being included in other libraries referenced
on the compilation line.  I believe the issue is that the inclusion
of libsmumps.so was (due to my original patch) causing
libmumps_common.so to be automatically loaded, but since libpords.so
was not also required, the error was occuring.  I have added libpords.so
along with libmumps_common.so to be explicit dependencies of
libsmumps.so, etc., which seems to resolve the issue.