libsodium: Fix build on ppc64le (#21256)
On ppc64le and aarch64, Spack tries to execute any "config.guess" and
"config.sub" scripts it finds in the source package.

However, in the libsodium tarball, these files are present but not
executable.  This causes the following error when trying to install
libsodium with spack:

    Error: RuntimeError: Failed to find suitable substitutes for config.sub, config.guess

Fix this by chmod-ing the scripts in the patch() function of libsodium.