hdf5: fix compiler detection in flag_handler (#24451)
The original implementation of `flag_handler` searched the
`self.compiler.cc` string for `clang` or `gcc` in order to add a flag
for those compilers.  This approach fails when using a spack-installed
compiler that was itself built with gcc or clang, as those strings will
appear in the fully-qualified compiler executable paths.  This commit
switches to searching for `%gcc` or `%clang` in `self.spec`.

Co-authored-by: Paul Henning <phenning@lanl.gov>