ArrayFire: new version and update CUDA usage (#29988)
* Set CUDA architectures in ArrayFire based on cuda_arch

The cuda_arch flag was not recognized by the ArrayFire package and
therefore any setting was not respected. This commit adds the appropriate
cmake flags if cuda_arch is specified. If no cuda_arch is specified,
then the flag is set to "Auto" which checks the installed compute
architectures on the build system.

* ArrayFire only requires boost headers to build. Update version to 1.75

ArrayFire only requires boost headers at build time. This commit also
updates the version to 1.75 to avoid some errors in Boost Compute

* Disable tests in ArrayFire by default

* Add support for ArrayFire v3.8.1

* Add maintainer for ArrayFire package

* Remove test variant from ArrayFire. Use comprehensions

* Reduce boost requirement in ArrayFire

* Address cuda_arch suggestions

* Add commit hashes to Release versions of ArrayFire

* Fix style issues in ArrayFire package