pgplot: install rbg.txt and change PGPLOT_DIR (#32775)
The file rbg.txt is needed for many PGPLOT application, such as the MESA
stellar evolution code. This change installs the file to the PGPLOT_DIR,
where the library expects it.

PGPLOT_DIR was previously set to the prefix itself, which is an odd
place to install rgb.txt. This commit changes it to lib/pgplot5,
following the convention used by Debian.

Co-authored-by: Philipp Edelmann <edelmann@fs.tum.de>