Move python glue logic - 2nd try (#50616)
Redo of #50605 (Move external python glue logic into core) 

This concerns logic that attaches a `python` dependency to some external
packages (because they expect an instance of `python` in their DAG).
#50605 was intended to be a simple relocation of logic from the package
repo to the core.

That PR caused a failure because it was attempting to attach python to a
bigger set of dependents: originally this logic only ran for packages
that were a subclass of `PythonPackage`, and that PR made it run for all
packages that `extends(python)`; notably that included LLVM, and in
particular this occurred inside of an environment that set `unify: false`;
attaching Python can involve looking for an instance of Python on the
system, and normal external detection runs with a process pool, which is
prohibited in the context of `concretize_separately` (which itself uses a
process pool with daemon processes).

This PR reapplies the logic from #50605 with two additional changes:

* Match the previous inclusion/exclusion semantics by only attaching
  Python to external packages that are a subclass of `PythonPackage`
* Update external Python search to run within the current process
  (so that it can run as part of `concretize_separately`)