mgard: Add configuration for compiling mgard with hip (#44)
This also provides patches for properly compiling MGARD with hip.
These include fixing compile errors for a bad types in templates that
was not caught by compilers that do not check non-instantiated
templates. It also has a patch to properly configure the hipcub library.
These patches will be part of MGARD after version 1.5.