intel-oneapi: enable determining openmp runtime (#468)
- intel-oneapi-mkl: enable threads!=none (see https://github.com/spack/spack-packages/pull/16 and https://github.com/spack/spack-packages/pull/401).
- intel-oneapi-mkl: made TBB-dependency conditional.
- Fixed missing base class (IntelOneApiStaticLibraryList).

Recent changes in Spack seem to be related to the issue. For the time being (prior to Spack 2.0), a resolution or workaround is needed. Having a workaround like depends_on("c", type="build") at the level of oneapi.py seems reasonable (intel-oneapi-mkl package can avoid a direct dependency on the workaround).