paraview: use viskores in newer versions
Starting with ParaView 6.0, the VTK-m accelerator library changed
to Viskores. Update the options to use the new accelerator library.

The approach has changed slightly. Earlier there was a use_vtkm
option to specify whether VTK-m should be compiled, and it always
compiled the version provided with ParaView. However, the API
for Viskores is stable, so use the external viskores package.

Co-authored-by: Ryan Krattiger <80296582+kwryankrattiger@users.noreply.github.com>