CI: Run release tags as noop jobs
Release tag copy can be done using other infrastructure tricks that are
not subject to non-determinism in concretization. Due to the requirement
for protected publish, something has to run and complete in CI for the
protected publish job, and the subsequent cache indexing jobs, to run.
This forces all generate steps to run noop jobs and nothing else.

Signed-off-by: Ryan Krattiger <ryan.krattiger@kitware.com>