formatting: make `ruff format` almost pass
This is intended to make moving to `ruff` easy. It is the result of
running:

```console
ruff format .
black .
```

which formats everything `ruff`'s way, then formats the parts `black`
doesn't like `black`'s way.  There are 11 files remaining on which they
disagree.

The goal here is to get the repository mostly formatted before moving
to `ruff`. Mostly, this eliminates unnecessary string concatenation in
packages, which makes them more readable and less prone to error, anyway.

Signed-off-by: Todd Gamblin <tgamblin@llnl.gov>