ci: use `python -m pip install` instead of `pip install` (#885)
The windows audits (at least) has started to fail with an error if we
invoke pip as `pip install` and not with `python -m pip install`.

> do you know why pip fails?

I initially thought this was a stricter warning in newer python, but
the python version is the same on ubuntu and windows. 

I don't know why it only started failing now, but apparently it's
failing specifically on windows because using:

```
pip install 
```

instead of:

```
python -m pip install
```

means `pip.exe` can't be upgraded because `pip.exe` is already
running (something Linux is fine with). Running `python -m pip`
circumvents this and allows pip to modify "itself".

-[x] Update CI to use `python -m pip install`

Signed-off-by: Todd Gamblin <tgamblin@llnl.gov>