network-direct: remove MC and add platform toolset definiton
On CI machines (but not locally) network direct is failing due to an attempt to load a default platform toolset for Windows 10.
Instead, add the CLI argument to specify the platform toolset for our selected version of MSVC.

For whatever reason, specifying the platform toolset causes message compiler detection to fail silently.
Thanks to Spack being Spack, we can guaruntee that we have an available MC, so we can replace MC references with the equivalent command line in the MS build system.