curl: add 8.14.1 and 8.15.0 (#739)
Update to the latest 8.14.1 and 8.15.0 releases.

`cmake` has some issues with `curl` 8.14.1 and `8.15` -- specifically,
`cmake` 3.11:3.31.6` need a patch, and older versions won't work
with these `curl` versions.

`curl` 8.15 removes support for `secure_transport` TLS on macOS,
which was our default TLS version. So, we need to make it rely on
`openssl`. This is not too disruptive, as everyone is either installing
`openssl` or pointing at an external one anyway.

---------

Signed-off-by: Todd Gamblin <tgamblin@llnl.gov>
Co-authored-by: John Parent <john.parent@kitware.com>