ruff: check for undefined symbols in packages (#924)
Playing around with `ruff` a bit more.

`ruff` can have a list of `builtins` in `pyproject.toml` that lets us
define which symbols can be used without being imported. I added
the spack package API, and any globals we're using in packages to
this. This enables us to stop ignoring F405 and to get undefined
symbol checks working on packages.

I think we can use this to check for Spack API violations and to curtail
how many magic globals get into `builtin`.  I think we can gradually
whittle them down now that they are enumerated in one place and
checkable with `ruff`.

- [x] update `[tool.ruff]` in `pyproject.toml` with `builtins` for
       everything in the package API plus globals from `build_environment`
       and various packages and build systems.
- [x] fix imports in tests
- [x] mark undefined symbols in tests with `noqa`
- [x] fix undefined symbols in all remaining packages in `builtin`

---------

Signed-off-by: Todd Gamblin <tgamblin@llnl.gov>
Co-authored-by: Harmen Stoppels <me@harmenstoppels.nl>