Fix external Intel classic compiler (#981)
With this change, one can `spack compiler add` an instance of
`intel-oneapi-compilers-classic` and use it (before, it was only
working when installed with Spack). This omits run-time env setup
that would be performed by an `intel-oneapi-compilers` dependency.
In testing, it did not seem necessary to do this in order to use
the compiler (at least for Intel 19 and 2021).

---------

Signed-off-by: Peter Scheibel <scheibel1@llnl.gov>