noop for all of the copy-only type jobs
Since the non-determinism effects all copy-only type jobs, and this is
only supposed to allow forcing running protected-publish after a
snapshot is done.

Currently the process is to create snapshots using a local script, and
then pushing the corresponding tag which will trigger this noop job
which will signal the rest of the workflow to run.