openmx: fix build with latest gcc, aocc and intel-oneapi (#681)
* openmx: add explicit blas and lapack dependencies

We are already using spec["lapack"].libs and spec["blas"].libs so this
is needed to not error with mkl (builds with netlib-scalapack or amd
scalapack work fine surprisingly).

* openmx: fix build with latest gcc and intel compilers

-fcommon is now needed also with oneapi and won't hurt for other
llvm-based compilers, so move it out of the gcc-only block.

Allow OpenMP build also for Fortran sources, which is crucial to enable
the threading in built-in ELPA, scaling is quite bad otherwise. Also
this is consistent with the upstream flags.

Also adds a bunch of flags for linking Fortran OpenMP code with C linker.