meson: deprecate superseded point releases
Per usual, assume that a feature release line with multiple bugfix
releases does not need every bugfix release.

As a special example, meson 1.8.* was a story in "we released a daring
refactor and then had repeated regressions fixed over the course of
several bugfix releases". You *really* do not want to be using 1.8.2 at
this point.

1.8.4 is "probably safe", as that was the specific version to include a
particularly large series of bugfixes, but... why chance it if users can
be encouraged to prefer 1.8.5 instead?