Fix OpenMP support in libCEED/MFEM
As I was working on the Palace package.py, I noticed that our tests were
failing when the OpenMP variant is
enabled (https://github.com/awslabs/palace/issues/557).

I tracked this down to the libCEED package not building with the
OPENMP=1 option with OpenMP is enabled.

I added a new `+openmp` variant to libCEED to handle this case.

I looked at the dependants and found that mfem should also be updated to
propagate the variant.

In this, I also noticed that MFEM describes the `threadsafe` variant
with
```
Enable thread safe features. Required for OpenMP. May cause minor
performance issues.
```

But does not activate the variant when `+openmp`. I now turn this
variant on when mfem is being compiled with OpenMP.