Fix LuaBuilder.unpack for non-existant archive_file
`LuaBuilder.unpack` attempts to create a Path object of
`pkg.stage.archive_file` without checking whether or not the value is
`None`. When a Lua package, such as `lua-ffi`, is installed from a git
commit, there is no archive file to unpack and an exception is thrown.
Reference error shown in
https://github.com/spack/spack-packages/issues/2518