Fix static builds in libceed
`libceed` did not have the `shared` variant and always built as a shared
library and it was not possible to compile `libceed` as part of a static
build. The reasons for this are two:

1. `libceed` has a build flag `STATIC `that has to turned on to build
`libceed` as a static library
2. the dependencies of `libceed` were not compiled as static libraries.

This resulted in errors like:
```
relocation R_X86_64_PC32 against symbol `magma2lapack_constants' can not be used when making a shared object; recompile with -fPIC
```
when building libceed as part of a static build.