Add libxcrypt dependency for screen
`screen` is linked to `libcrypt` for versions 3.9 and higher. In
`screen` 3.7, the `configure` script only checks for `libcrypt.a` in
`/usr` and `/usr/build` and does not use `-lcrypt` if neither are found
(thus I conclude that it's optional). It is not optional for newer
versions and the build fails if `libcrypt` isn't found. Thus for
`screen` 3.9 and newer, we add a dependency on the `libxcrypt` package.