Add cxx dependency for hugo extended edition
Per the `hugo`
[README](https://github.com/gohugoio/hugo?tab=readme-ov-file#build-from-source),
A compiler is needed when building the extended edition (using the
`+extended` variant), but the package is missing this dependency. I've
found that when building `hugo+extended`, a build dependency on `cxx` is
sufficient to ensure that the binary is successfully linked to libstdc++
and libgcc_s from the `gcc-runtime` package, and the binary is
successfully relocated when installed from buildcache.