Add procps dependency to lmod
Way back in Lmod 6.0.12, with [this commit](https://github.com/TACC/Lmod/commit/0ddb7bd2fdc462915205bf4855b18b9229a57548), Lmod added a dependency on
the `ps` command. I encountered a failure to build Lmod when building in
a container that did not have `ps` available. Thus, we add a dependency
on `procps`.