Remove samtools dependency
It turns out that `preseq` does not have an external `samtools`
dependency at all, and the Spack-built `samtools` is not sufficient in
any case (as it merely builds and isntalls the samtools binaries and not
the samtools source code). `preseq` 2 requires the raw `samtools`
source, and it includes it in its release tarball.