clingo-bootstrap: optimize hash table
In clingo@spack the hottest code path is `hash_set.hh`. It uses a bit of
a non-standard hash table implementation, where the container has a
capacity that's a prime number. That's probably to make hash collisions
unlikely, but in practice the hash function is fine, and it's more
efficient to use powers of two so that computing the slot is fast.

Also precomputes the threshold under which growing/rehashing is needed.

Signed-off-by: Harmen Stoppels <me@harmenstoppels.nl>